New Updated AddStudetForm
Created a function name CheckIfTextBoxesEmpty checking if one of the textboxes in form are empty and if so pops a message.

Also made so that you cant input chars into phone and id text boxes.

Also made it so that the numbers in grades are only between 0 to 100 and 777 only